// .integration-src/openvenice-master/src/types/venice.ts
export interface ImageConstraints {
  promptCharacterLimit?: number
  aspectRatios?: string[]
  defaultAspectRatio?: string
  resolutions?: string[]
  defaultResolution?: string
  steps?: { default: number; max: number }
  widthHeightDivisor?: number
}

export interface VideoConstraints {
  model_type: 'text-to-video' | 'image-to-video'
  aspect_ratios: string[]
  resolutions: string[]
  durations: string[]
  audio: boolean
  audio_configurable: boolean
  audio_input: boolean
  video_input: boolean
}

export interface ModelCapabilities {
  optimizedForCode?: boolean
  quantization?: string
  supportsAudioInput?: boolean
  supportsFunctionCalling?: boolean
  supportsLogProbs?: boolean
  supportsMultipleImages?: boolean
  supportsReasoning?: boolean
  supportsReasoningEffort?: boolean
  supportsResponseSchema?: boolean
  supportsTeeAttestation?: boolean
  supportsE2EE?: boolean
  supportsVideoInput?: boolean
  supportsVision?: boolean
  supportsWebSearch?: boolean
  supportsXSearch?: boolean
}

export type ModelTrait =
  | 'default'
  | 'most_intelligent'
  | 'most_uncensored'
  | 'function_calling_default'
  | 'default_reasoning'
  | 'default_code'
  | 'default_vision'

export interface VeniceModel {
  id: string
  object: string
  created: number
  owned_by: string
  model_spec?: {
    availableContextTokens?: number
    maxCompletionTokens?: number
    capabilities?: ModelCapabilities
    traits?: ModelTrait[]
    offline?: boolean
    name?: string
    description?: string
    constraints?: VideoConstraints | ImageConstraints
    model_sets?: string[]
    pricing?: {
      input?: { usd?: number }
      output?: { usd?: number }
    }
  }
}

export interface ModelsResponse {
  object: string
  data: VeniceModel[]
}